Implementation of a lock system for feature parameters

+++ vx_library/vx_shell/features/parameters/parameters.py
@@ -1,54 +1,26 @@
-import json
-from typing import Literal, Optional, Dict
-from pydantic import BaseModel, validator
+import os, json
+from typing import Literal, Optional, Dict, Any
+from pydantic import BaseModel
+from .models import RootFeatureParams, UserFeatureParams, FrameParams
+from ...globals import ROOT_CONFIG_DIRECTORY, USER_CONFIG_DIRECTORY
+
+ROOT_PARAMS_DIRECTORY = f"{ROOT_CONFIG_DIRECTORY}/features"
+USER_PARAMS_DIRECTORY = f"{USER_CONFIG_DIRECTORY}/features"
 
 
-def read_json(path: str) -> dict:
-    with open(path, "r", encoding="utf-8") as file:
-        return json.load(file)
+def read_json(file_path: str) -> dict | None:
+    if os.path.exists(file_path):
+        with open(file_path, "r", encoding="utf-8") as file:
+            return json.load(file)
 
 
-def write_json(path: str, data: dict):
-    with open(path, "w", encoding="utf-8") as file:
-        json.dump(data, file, indent=4, ensure_ascii=False)
-
-
-LevelKeys = Literal["background", "bottom", "overlay", "top"]
-AnchorEdgeKeys = Literal["top", "right", "bottom", "left", "full"]
-AlignmentKeys = Literal["start", "center", "end", "stretch"]
-
-
-class MarginParams(BaseModel):
-    top: Optional[int] = None
-    right: Optional[int] = None
-    bottom: Optional[int] = None
-    left: Optional[int] = None
-
-
-class LayerFrameParams(BaseModel):
-    monitor_id: Optional[int] = None
-    auto_exclusive_zone: Optional[bool] = None
-    exclusive_zone: Optional[int] = None
-    level: Optional[LevelKeys] = None
-    anchor_edge: Optional[AnchorEdgeKeys] = None
-    alignment: Optional[AlignmentKeys] = None
-    margins: Optional[MarginParams] = None
-    width: Optional[int] = None
-    height: Optional[int] = None
-
-
-class FrameParams(BaseModel):
-    name: str
-    route: str
-    show_on_startup: Optional[bool] = None
-    layer_frame: Optional[LayerFrameParams] = None
-    multi_frame: Optional[bool] = None
-
-    @validator("multi_frame", pre=True)
-    def validate_multi_frame(cls, v, values):
-        if values.get("layer_frame") is not None and v is not None:
-            raise ValueError("A layer frame cannot be a multi frame")
-        return v
+def write_json(file_path: str, data: dict):
+    if os.path.exists(file_path):
+        with open(file_path, "w", encoding="utf-8") as file:
+            json.dump(data, file, indent=4, ensure_ascii=False)
+    else:
+        with open(file_path, "x", encoding="utf-8") as file:
+            json.dump(data, file, indent=4, ensure_ascii=False)
 
 
 class FeatureParams(BaseModel):
@@ -58,14 +30,99 @@ class FeatureParams(BaseModel):
     state: Optional[Dict[str, None | str | int | float | bool]] = None
 
     @staticmethod
-    def create(path: str):
-        data: Dict = read_json(path)
-        data["path"] = path
+    def create(file_name: str) -> dict:
+        def root_validator(root_value: Any | Literal["disable"] | None):
+            if root_value is not None:
+                if root_value != "disable":
+                    return root_value
+                else:
+                    return None
+            else:
+                return "__user__"
 
-        return FeatureParams(**data)
+        def get_params(root_data: Dict, user_data: Dict):
+            root_params = RootFeatureParams(**root_data).model_dump()
+            params = UserFeatureParams(**user_data).model_dump()
+
+            params["path"] = f"{USER_PARAMS_DIRECTORY}/{file_name}"
+            params["name"] = root_params["name"]
+
+            if not "frames" in params:
+                params["frames"] = {}
+
+            for root_frame_key, root_frame_item in root_params["frames"].items():
+                if not root_frame_key in params["frames"]:
+                    params["frames"][root_frame_key] = {}
+
+                params["frames"][root_frame_key]["name"] = root_frame_item["name"]
+                params["frames"][root_frame_key]["route"] = root_frame_item["route"]
+
+                show_on_startup = root_validator(root_frame_item["show_on_startup"])
+                if show_on_startup != "__user__":
+                    params["frames"][root_frame_key][
+                        "show_on_startup"
+                    ] = show_on_startup
+
+                multi_frame = root_validator(root_frame_item["multi_frame"])
+                if multi_frame != "__user__":
+                    params["frames"][root_frame_key]["multi_frame"] = multi_frame
+
+                layer_frame = root_validator(root_frame_item["layer_frame"])
+                if layer_frame != "__user__":
+                    if layer_frame is None:
+                        params["frames"][root_frame_key]["layer_frame"] = None
+                    else:
+                        for layer_frame_key, layer_frame_item in root_frame_item[
+                            "layer_frame"
+                        ].items():
+                            if layer_frame_key == "margins":
+                                margins = root_validator(layer_frame_item)
+                                if margins != "__user__":
+                                    if margins is None:
+                                        params["frames"][root_frame_key]["layer_frame"][
+                                            layer_frame_key
+                                        ] = None
+                                    else:
+                                        for margin_key, margin_item in root_frame_item[
+                                            "layer_frame"
+                                        ]["margins"].items():
+                                            value = root_validator(margin_item)
+                                            if value != "__user__":
+                                                params["frames"][root_frame_key][
+                                                    "layer_frame"
+                                                ]["margins"][margin_key] = value
+                            else:
+                                value = root_validator(layer_frame_item)
+                                if value != "__user__":
+                                    params["frames"][root_frame_key]["layer_frame"][
+                                        layer_frame_key
+                                    ] = value
+
+            state = root_validator(root_params["state"])
+            if state != "__user__":
+                params["state"] = state
+
+            return params
+
+        return FeatureParams(
+            **get_params(
+                read_json(f"{ROOT_PARAMS_DIRECTORY}/{file_name}"),
+                read_json(f"{USER_PARAMS_DIRECTORY}/{file_name}"),
+            )
+        )
 
     def save(self):
-        write_json(self.path, self.model_dump(exclude={"path"}, exclude_none=True))
+        frames = {}
+
+        for key in self.frames:
+            frames[key] = self.frames[key].model_dump(
+                exclude={"name", "route"}, exclude_none=True
+            )
+
+        data = self.model_dump(exclude={"path", "name", "frames"})
+        data["frames"] = frames
+
+        write_json(self.path, data)
 
     @property
     def single_frames(self) -> Dict[str, FrameParams]:
